requirements.txt
numpy==1.26.3
opencv-python==4.8.0.74
Pillow==10.1.0
scikit-learn==1.3.2
torch==2.0.0
transformers==4.30.0
